#6 refactor : 폴더명 수정
